+++ metadata.rb
@@ -4,7 +4,7 @@ maintainer_email  'oraclesunspar-at-gmail-dot-com'
 license           'MIT'
 description       'Provides LWRPs for installing and managing TeamSpeak 3 servers.'
 long_description  IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version           '0.2.0'
+version           '0.3.0'
 source_url        'https://github.com/Sunspar/ts3'
 issues_url        'https://github.com/Sunspar/ts3/issues'